requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
alembic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
asgiref
pydantic[email]
python-multipart
